nvs: support StorageIterate
This impliments `StorageIterate` for `EspNvs`. Note that while the
`Error` on iteration is setup to work, the actual implimentation in
esp-idf does not return any errors durring iteration (other than "Not
Found", indicating iteration is complete). As a result, `next()` will
always return `Some(Ok(...))` or `None` and never `Some(Err(...))`.